+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK7DZPAMGPS257Z5P1DZWM6B"
+        "01KK8GKB59SNWQZT86TRZVNYVS"
     )
     r"""The unique identifier of the data source"""