+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK7DZPAK8CSR8NG4MKJGPSDH"
+        "01KK8GKB58V11WM75XHZDM7ARQ"
     )
     r"""The unique identifier of the data source"""