+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK7DZPAPBA4A3RR22DX38716"
+        "01KK8GKB5BKVACEBCGJYJCTZ3E"
     )
     r"""The unique identifier of the data source"""