+++ packages/orq-rc/RELEASES.md
@@ -6369,3 +6369,13 @@ Based on:
 - [python v4.5.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.39 - packages/orq-rc
+
+## 2026-03-09 05:22:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.749.1 (2.855.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.40 - packages/orq-rc